requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
dj-database-url==2.3.0
Django==4.2.16
django-cors-headers==4.6.0
django-environ==0.11.2
djangorestframework==3.15.2
gunicorn==23.0.0
idna==3.10
mailersend==0.5.8
packaging==24.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.2
typing_extensions==4.12.2
urllib3==2.2.3
